+++ src/lib.rs
@@ -227,6 +227,64 @@ impl<T: PartialOrd> Range<T> {
     }
 }
 
+/// Range filter.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+pub struct DateRange {
+    /// The filter's lower bound.
+    #[serde(rename = "from")]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(serialize_with = "serde::rfc3339_micros_opt")]
+    pub lower: Option<DateTime<Utc>>,
+
+    /// The filter's upper bound.
+    #[serde(rename = "to")]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(serialize_with = "serde::rfc3339_micros_opt")]
+    pub upper: Option<DateTime<Utc>>,
+}
+
+impl DateRange {
+    /// Create a new range filter.
+    #[must_use]
+    pub const fn new() -> Self {
+        Self {
+            lower: None,
+            upper: None,
+        }
+    }
+
+    /// Specify a 'greater-than' lower bound for the filter.
+    #[must_use]
+    pub const fn gt(mut self, gt: DateTime<Utc>) -> Self {
+        self.lower = Some(gt);
+        self
+    }
+
+    /// Specify a 'less-than' upper bound for the filter.
+    #[must_use]
+    pub const fn lt(mut self, lt: DateTime<Utc>) -> Self {
+        self.upper = Some(lt);
+        self
+    }
+
+    /// Check if the range contains the value.
+    #[must_use]
+    pub fn contains(&self, value: &DateTime<Utc>) -> bool {
+        if let Some(lower) = &self.lower {
+            if value < lower {
+                return false;
+            }
+        }
+        if let Some(upper) = &self.upper {
+            if value > upper {
+                return false;
+            }
+        }
+
+        true
+    }
+}
+
 // Custom serialization functions.
 mod serde {
     use chrono::{DateTime, SecondsFormat, Utc};
