+++ tests/records_query.rs
@@ -5,9 +5,9 @@ use dwn_node::clients::protocols::ConfigureBuilder;
 use dwn_node::clients::records::{Data, ProtocolBuilder, QueryBuilder, WriteBuilder};
 use dwn_node::data::{DataStream, MAX_ENCODED_SIZE};
 use dwn_node::protocols::Definition;
-use dwn_node::records::{DateRange, RecordsFilter, Sort};
+use dwn_node::records::{RecordsFilter, Sort};
 use dwn_node::store::Pagination;
-use dwn_node::{Error, Message, Range, authorization, endpoint};
+use dwn_node::{DateRange, Error, Message, Range, authorization, endpoint};
 use dwn_test::key_store::{self, ALICE_DID, BOB_DID, CAROL_DID};
 use dwn_test::provider::ProviderImpl;
 use http::StatusCode;
